Module Math.Monomial

Abelian group of abstract monomial in arbitrary number of coordinates.

A monomial represents \(x_0^{a_0}··· x_n^{a_n}\) where \(x_j\) are formal indeterminates and \(a_j\) are integers (negative exponents are allowed). The group operation is the multiplication of monomials.

This group is isomorphic to the group of sequences of integers with finite support (the sequences of exponents \([a_0,a_1,...]\)), where the group operation is now the addition. Hence the API exposes both the multiplicative and the additive notation.

module type S = sig ... end

module Generic : S with type names = int → string option

Generic implementation of monomials