Module Monomial.Generic

Generic implementation of monomials

The default Monomial.S.names is the map i => \(x_i\).

include AbelianGroup

type t

val zero : t

Neutral element for addition.

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val equal : t → t → bool

val (==) : t → t → bool

val one : t

The constant mononial 1; same as zero, in additive notation.

val degree : t → int

Total degree

val support : t → Iset.t

support m is the list of indices of m with non-zero exponents

val imax : t → int option

Max index i of variables \(x_i\) in the mononial, or None.

val exponent : int → t → int

exponent i m is the exponent of the i-eth variable.

val xi : int → t

xi i is the monomial \(x_i\)

val xin : int → int → t

xin i n is the monomial \(x_i^n\)

val mul : t → t → t

val (*) : t → t → t

multiplication of two monomials; same as addition of exponents.

val div : t → t → t

val (/) : t → t → t

div m1 m2 is the division of m1 by m2. This can produce negative exponents. Same as substraction of exponents.

val imul : int → t → t

imul i m is m muliplied by \(x_i\)

val of_list : (int * int) list → t

Create a monomial \(x_{i_1}^{a_1}x_{i_2}^{a_2}\dots\) from a list of the form [(i1,a1); (i2,a2), ...]. If the same index i appears several times, only the last one will be taken into account.

val to_list : t → (int * int) list

module Compare : sig ... end

type names = int → string option

A variable of type names is used to print out the variables. The precise type depends on the implementation.

val default_names : names Pervasives.ref

val to_tex : ?⁠names:names → t → string

val of_tex : string → t