Module Math.Monomial1

A simpler implementation for Monomials in 1 variable \(x_0\)

include Monomial.S with type names = string

include AbelianGroup

type t

val zero : t

Neutral element for addition.

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val equal : t → t → bool

val (==) : t → t → bool

val one : t

The constant mononial 1; same as zero, in additive notation.

val degree : t → int

Total degree

val support : t → Iset.t

support m is the list of indices of m with non-zero exponents

val exponent : int → t → int

exponent i m is the exponent of the i-eth variable.

val xi : int → t

xi i is the monomial \(x_i\)

val xin : int → int → t

xin i n is the monomial \(x_i^n\)

val mul : t → t → t

val (*) : t → t → t

multiplication of two monomials; same as addition of exponents.

val div : t → t → t

val (/) : t → t → t

div m1 m2 is the division of m1 by m2. This can produce negative exponents. Same as substraction of exponents.

val imul : int → t → t

imul i m is m muliplied by \(x_i\)

val of_list : (int * int) list → t

Create a monomial \(x_{i_1}^{a_1}x_{i_2}^{a_2}\dots\) from a list of the form [(i1,a1); (i2,a2), ...]. If the same index i appears several times, only the last one will be taken into account.

val to_list : t → (int * int) list

module Compare : sig ... end

type names = string

A variable of type names is used to print out the variables. The precise type depends on the implementation.

val to_tex : ?⁠names:names → t → string

val of_tex : string → t

val x : t

The \(x\) monomial

val xn : int → t

The \(x^n\) monomial

val to_generic : int → t → Monomial.Generic.t

to_generic i m transforms the 1D monomial m into a generic monomial where the original variable is replaced by \(x_i\).

val set_default_name : names → unit