Parameter Generic.1-R

include Rng

include AbelianGroup

type t

val zero : t

Neutral element for addition.

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val equal : t → t → bool

val (==) : t → t → bool

val mul : t → t → t

val (*) : t → t → t

Ring multiplication.

val pow : int → t → t

pow n x is x raised to the power n.

val one : t

Ring unit element.

val mone : t

minus one.

val of_int : int → t

type names

If the type names remains private, it means it has no use in the current module implementation.

val to_tex : ?⁠names:names → t → string

val of_tex : string → t