Module Math.Polynomial1

Polynomials in one variable.

A module of type Polynomial1 can be used as a Polynomial.

The default Monomial.S.names is the string "x"

module type S = sig ... end

module Make : functor (R : Ring) → S with type scalar = R.t

Polynomials over polynomials

Since polynomials form a ring, one can construct polynomials in two variables as polynomials in one variable over the ring of polynomials in one variable. Namely, one can write fo instance:

module RatPoly2 = Polynomial1.Make(RatPoly1)

Elements of RatPoly2 are now polynomials in one variable whose coefficients are polynomials in one variable with rational coefficients; they can be viewed as polynomials in two variables with rational coefficients. In order to distinguish the two variables in LaTeX output, one should do for instance RatPoly1.set_default_name "t".