Module Polynomial1.Make

Parameters

R : Ring

Signature

include Polynomial(Monomial1).S with type names = string

type scalar

Type of the scalar Ring.

type t

val zero : t

Neutral element for addition.

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val equal : t → t → bool

val (==) : t → t → bool

val mul : t → t → t

val (*) : t → t → t

Ring multiplication.

val pow : int → t → t

pow n x is x raised to the power n.

val one : t

Ring unit element.

val mone : t

minus one.

val of_int : int → t

type names = string

val to_tex : ?⁠names:names → t → string

val of_tex : string → t

val of_scalar : scalar → t

of_scalar s is scal_mul s one

val scal_mul : scalar → t → t

type monomial = Monomial1.t

val degree : t → int option

Degree of a non-zero polynomial, or None for the zero polynomial.

val idegree : int → t → int option

Degree in the i-eth variable.

val imax : t → int option

Maximum index i of variables \(x_i\) appearing in the polynomial. If n = imax p then p can be seen as a polynomial in \((x_0,\dots,x_n)\). Returns None if p is zero.

val const : scalar → t

Constant polynomials

val xi : int → t

xi i is the polynomial \(x_i\)

val of_monomial : monomial → t

Return the polynomial consisting of the given monomial multiplied by the unit ring element.

val add_monomial : monomial → scalar → t → t

add_monomial c m p adds the monomial m with coefficient c to the polynomial p

val partial : int → t → t

partial i p is the partial derivative of p with respect to the i-eth variable.

val diff : monomial → t → t

diff monomial : t -> t is the differential operator obtained by replacing the ieth variable in the monomial by the ieth derivative.

val map : (scalar → scalar) → t → t

Apply a function on the coefficients of the polynomial.

val of_list : (scalar * monomial) list → t

Create a polynomial by adding all monomials with their given coefficients.

val to_list : t → (scalar * monomial) list

val default_names : names Pervasives.ref

type generic

val x : t

The \(x\) polynomial

val of_array : scalar array → t

For instance of_array [| 3; 4; 5|] is the polynomial \(3 + 4x + 5x^2\).

val to_generic : int → t → generic

to_generic i p transforms the 1D polyomial p into a generic polyomial where the original variable is replaced by \(x_i\).

val set_default_name : string → unit

Set the default string used to convert the variable to LaTeX code. Default is "x".