Module Math.Weyl

Semiclassical Weyl Algebra

A semiclassical Weyl algebra is a formal deformation of a Poisson algebra. It is a polynomial algebra with variables \((\hbar, q_1, p_1, q_2, p_2, \dots)\). The degree in the hbar variable (\(\hbar\)) is 2. It is a Lie algebra (LieAlg) with the so-called Moyal bracket.
\[[ f,g ] (q,p , \hbar ) = 2 \sinh \Bigl( \frac{\hbar}{2i} \Pi \Bigr)\bigl( f ( q, p, \hbar ) g ( q', p', \hbar ) \bigr) \Bigr|_{q=q',\atop p=p'}\]
with the bi-Poisson operator \[ \Pi = \sum_{j=1}^n \partial_{p_j} \partial_{q'_j} - \partial_{q_j} \partial_{p'_j} .\]

The sign convention here is opposite to the formula in the wikipedia page. At first order in ħ, the Moyal bracket reduces to the Poisson bracket:
\[[ f,g ] = \frac\hbar{i}\{f,g\} + O(\hbar^3) \]

module type Poisson = sig ... end

A Poisson algebra is both an Algebra and a LieAlg. The Lie bracket is the Poisson bracket \(\{f,g\}\). We should have \(\{f,gh\} = g\{f,h\} + \{f,g\}h\) (but of course this is not guaranteed by the signature).

module Monomial : sig ... end

Monomials in \((\hbar, q_1, p_1, q_2, p_2, \dots)\).

module Make : functor (F : Field) → sig ... end

Construct a Weyl algebra over the field F.