Module Weyl.Make

Construct a Weyl algebra over the field F.

A module constructed by this function can be used as a Poisson.

Parameters

F : Field

Signature

include Polynomial(Monomial).S with type scalar = F.t

type scalar = F.t

type t

val zero : t

Neutral element for addition.

val is_zero : t → bool

val neg : t → t

val add : t → t → t

val (+) : t → t → t

val sub : t → t → t

val (-) : t → t → t

val equal : t → t → bool

val (==) : t → t → bool

val mul : t → t → t

val (*) : t → t → t

Ring multiplication.

val pow : int → t → t

pow n x is x raised to the power n.

val one : t

Ring unit element.

val mone : t

minus one.

val of_int : int → t

type names

If the type names remains private, it means it has no use in the current module implementation.

val to_tex : ?⁠names:names → t → string

val of_tex : string → t

val of_scalar : scalar → t

of_scalar s is scal_mul s one

val scal_mul : scalar → t → t

type monomial = Monomial.t

val degree : t → int option

Degree of a non-zero polynomial, or None for the zero polynomial.

val idegree : int → t → int option

Degree in the i-eth variable.

val imax : t → int option

Maximum index i of variables \(x_i\) appearing in the polynomial. If n = imax p then p can be seen as a polynomial in \((x_0,\dots,x_n)\). Returns None if p is zero or one.

val const : scalar → t

Constant polynomials

val xi : int → t

xi i is the polynomial \(x_i\)

val of_monomial : monomial → t

Return the polynomial consisting of the given monomial multiplied by the unit ring element.

val add_monomial : monomial → scalar → t → t

add_monomial c m p adds the monomial m with coefficient c to the polynomial p

val partial : int → t → t

partial i p is the partial derivative of p with respect to the i-eth variable.

val diff : monomial → t → t

diff monomial : t -> t is the differential operator obtained by replacing the ieth variable in the monomial by the ieth derivative.

val map : (scalar → scalar) → t → t

Apply a function on the coefficients of the polynomial.

val of_list : (scalar * monomial) list → t

Create a polynomial by adding all monomials with their given coefficients.

val to_list : t → (scalar * monomial) list

val default_names : names Pervasives.ref

val hbar : t

val qi : int → t

The \(q_i\) coordinate.

val pi : int → t

The \(p_i\) coordinate.

val poisson : t → t → t

val poisson2 : t → t → t

The Poisson bracket: \[ \{f,g\} = \sum_{i=1}^n \left(\frac{\partial f}{\partial p_i}\frac{\partial g}{\partial q_i} - \frac{\partial f}{\partial q_i}\frac{\partial g}{\partial p_i}\right). \]

val moyal : t → t → t

moyal f g is the Weyl deformation of the Poisson bracket. It is equal to \(i/\hbar\) times the Moyal bracket of f and g (not the Moyal product):
\[ \frac{i}{\hbar}[ f,g ] (q,p , \hbar ) = \Delta^* \sum_{k\geq 0} (-1)^{k} \frac{\hbar^{2k}}{2^{2k} (2k+1)!} \Pi^{2k+1} (f\otimes g) = \{f,g\} + O(\hbar^3) , \]
where \(\Pi\) is the bi-Poisson operator and \(\Delta^*\) is the contraction operator (PolyTensor.contract). See also the wikipedia page, but the sign convention there is opposite.

We don't define the Moyal star product here, because it requires complex coefficients. See the CWeyl module.