Module Math.Weyl

Semiclassical Weyl Algebra

A semiclassical Weyl algebra is a polynomial algebra with variables \((\hbar, q_1, p_1, q_2, p_2, \dots)\). The degree in the hbar variable (\(\hbar\)) is 2.

It is a Lie algebra (LieAlg) with the so-called Moyal bracket.

At first order in ħ, the Moyal bracket reduces to the Poisson bracket. The sign convention here is opposite to the formula in the wikipedia page.

module Monomial : sig ... end

module Make : functor (R : Ring) → sig ... end

Construct a Weyl algebra over the coefficient ring R.